% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{aipw_cov_bal_gt}
\alias{aipw_cov_bal_gt}
\title{aipw_cov_bal_gt}
\usage{
aipw_cov_bal_gt(g, tp, decomposed_aipw, X)
}
\arguments{
\item{g}{the group for which the balance statistics will be computed}

\item{tp}{the time period for which the balance statistics will be computed}

\item{decomposed_aipw}{a \code{decomposed_aipw} object that comes from
running \code{implicit_aipw_weights} in a previous step.}

\item{X}{a data frame that contains the mean of each extra balance variable
that was passed into the function \code{twfe_cov_bal}.}
}
\value{
a data frame that contains the balance statistics for each
covariate that was passed into the \code{extra_balance_vars_formula} argument
in \code{aipw_cov_bal}.
}
\description{
a (mainly internal) function to compute balance statistics
using the implicit weights from an AIPW regression for a particular
group \code{g} and time period \code{tp}.  This function is called internally
by \code{aipw_cov_bal}.
}
