% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{all_twfe_weights}
\alias{all_twfe_weights}
\title{all_twfe_weights}
\usage{
all_twfe_weights(
  yname,
  tname,
  idname,
  gname,
  xformla = NULL,
  xname = NULL,
  data
)
}
\arguments{
\item{yname}{The name of the outcome variable}

\item{tname}{The name of the column containing the time periods}

\item{idname}{The individual (cross-sectional unit) id name}

\item{gname}{The name of the variable in \code{data} that
contains the first period when a particular observation is treated.
This should be a positive number for all observations in treated groups.
It defines which "group" a unit belongs to.  It should be 0 for units
in the untreated group.}

\item{xformla}{formula for covariates that are included in the
TWFE regression.}

\item{xname}{by default, \code{twfe_weights_gt} applies the
implicit TWFE regression weights to the outcome.  In
cases where the researcher would like to apply the
implicit regression weights to some other variable, the
name of this variable can be passed in through this
argument.

This is typically used as a way to check how well
the TWFE regression balances levels of time-varying
covariates, time-invariant covariates, or other variables
not included in the model.  This is often useful
as a diagnostic to assess how sensitive the TWFE regression
is to violations of linearity.

@return a gt_weights object
@export}

\item{data}{The name of the data.frame that contains the data}
}
\value{
a \code{decomposed_twfe} object
}
\description{
a function to compute implicit TWFE weights
}
