% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights.R
\name{att_simple_weights}
\alias{att_simple_weights}
\title{att_simple_weights}
\usage{
att_simple_weights(attgt_res, w = rep(1, nrow(attgt_res$DIDparams$data)))
}
\arguments{
\item{attgt_res}{Results from did::att_gt function}

\item{w}{Optional external unit-specific weights}
}
\value{
a data.frame containing att_simple weights
}
\description{
A function to compute weights on ATT(g,t) to deliver
ATT^{simple} as discussed in Callaway and Sant'Anna (2021)
}
