% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights.R
\name{attO_weights}
\alias{attO_weights}
\title{attO_weights}
\usage{
attO_weights(attgt_res, w = rep(1, nrow(attgt_res$DIDparams$data)))
}
\arguments{
\item{attgt_res}{Results from did::att_gt function}

\item{w}{Optional external unit-specific weights}
}
\value{
a data.frame containing attO weights
}
\description{
A function to compute weights on ATT(g,t) to deliver
ATT^O as discussed in Callaway and Sant'Anna (2021)
}
