% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_post_lasso.R
\name{did_post_lasso}
\alias{did_post_lasso}
\title{did_post_lasso}
\usage{
did_post_lasso(
  y1,
  y0,
  D,
  covariates,
  pscore_covariates = covariates,
  i.weights = NULL,
  inffunc = TRUE,
  nfolds = 5
)
}
\arguments{
\item{y1}{Numeric vector of outcomes in the second time period}

\item{y0}{Numeric vector of outcomes in the first time period}

\item{D}{Numeric vector of treatment indicators}

\item{covariates}{Matrix of covariates}

\item{i.weights}{Numeric vector of sampling weights, defaault is NULL, so that observations
are equally weighted}

\item{inffunc}{Logical indicating whether to compute the influence function, default is TRUE}

\item{nfolds}{Number of folds for cross-validation, default is 5}

\item{alpha}{Elastic net mixing parameter, default is 1 for Lasso}
}
\value{
A list with the ATT estimate, standard error, selected variables for the outcome model,
selected variables for the propensity score model, and the influence function
}
\description{
Function to estimate the ATT using post-Lasso and compute the asymptotic variance with normalized weights
}
