% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_post_lasso.R
\name{did_post_lasso_ra}
\alias{did_post_lasso_ra}
\title{did_post_lasso_ra}
\usage{
did_post_lasso_ra(
  y1,
  y0,
  D,
  covariates,
  i.weights = NULL,
  inffunc = TRUE,
  alpha = 1
)
}
\description{
Function to estimate the ATT using post-Lasso and compute the asymptotic variance with normalized weights
that only uses regression adjustment.  This function requires stronger assumptions on sparsity that \code{did_post_lasso},
but it does not require the estimation of the propensity score, which may be infeasible in applications with
staggered treatment adoption and where the number of units per timing group is fairly small.
}
