% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_period_covs.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{effective_sample_size}
\usage{
effective_sample_size(
  est_weights,
  sampling_weights = rep(1, length(est_weights))
)
}
\arguments{
\item{est_weights}{a numeric vector of weights coming from an estimation procedure,
default is rep(1, length(n))}
}
\value{
a numeric value containing the effective sample size
}
\description{
a helper function to compute an effective sample size given
some estimation weights and sampling weights.  This function should be
called separately for the treated group and the untreated group.
}
