% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_period_covs.R
\name{frac_treated_extreme}
\alias{frac_treated_extreme}
\title{frac_treated_extreme}
\usage{
frac_treated_extreme(
  x,
  D,
  est_weights = rep(1, length(x)),
  sampling_weights = rep(1, length(x)),
  alpha = 0.05
)
}
\arguments{
\item{x}{a numeric vector}

\item{D}{a binary vector indicating treated and untreated group}

\item{alpha}{a numeric value between 0 and 1, default is 0.05}
}
\value{
a numeric value containing the fraction of treated units with extreme values
}
\description{
a helper function to compute the fraction of treated units that have
extreme values of a covariate relative to the untreated group.  The function
calculates the (alpha/2) and (1-alpha/2) quantiles of the covariate for the
untreated group and then computes the fraction of treated units that have
values of the covariate that are less than the (alpha/2) quantile or greater
than the (1-alpha/2) quantile.  This is a balance summary statistic that was
discussed in Imbens and Rubin (2015).
}
