% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_period_covs.R
\name{ggtwfe.two_period_covs_obj}
\alias{ggtwfe.two_period_covs_obj}
\title{a function to plot the balance statistics from the \code{two_period_covs} function}
\usage{
ggtwfe.two_period_covs_obj(two_period_covs_obj)
}
\arguments{
\item{two_period_covs_obj}{the object returned from the \code{two_period_covs} function}

\item{...}{additional arguments to pass to \code{ggplot2::geom_point}}
}
\description{
a function to plot the balance statistics from the \code{two_period_covs} function
}
