% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights.R
\name{ggtwfeweights.mp_weights_obj}
\alias{ggtwfeweights.mp_weights_obj}
\title{ggtwfeweights.mp_weights_obj}
\usage{
\method{ggtwfeweights}{mp_weights_obj}(x, ...)
}
\arguments{
\item{x}{An object of class mp_weights_obj}

\item{...}{Additional arguments to be passed to ggplot}
}
\description{
A function to plot weights on group-time average treatment effects
}
