% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ggtwfeweights}
\alias{ggtwfeweights}
\title{ggtwfeweights}
\usage{
ggtwfeweights(x, ...)
}
\arguments{
\item{x}{a \code{decomposed_twfe} object that comes from a
previous step of running either \code{twfe_cov_bal} or \code{aipw_cov_bal}.}

\item{...}{extra arguments, not used here}
}
\value{
a ggplot object
}
\description{
a generic function to plot balance statistics
}
