% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{gt_weights}
\alias{gt_weights}
\title{gt_weights}
\usage{
gt_weights(
  g,
  tp,
  treated = NULL,
  comparison = NULL,
  weights_treated,
  weights_comparison,
  weighted_outcome_treated = NULL,
  weighted_outcome_comparison = NULL,
  weighted_outcome_diff = NULL,
  alpha_weight
)
}
\arguments{
\item{g}{the current group for which the weights will be computed}

\item{tp}{the time period for which the weights will be computed}

\item{treated}{a vector of units that make up the effective treated group for group g
and period tp}

\item{comparison}{a vector of units that make up the effective comparison
group for group g and period tp}

\item{weights_treated}{a vector of weights for the treated group g in time
period tp}

\item{weights_comparison}{a vector of weights for the comparison group for
group g in time period tp}

\item{weighted_outcome_treated}{the mean outcome for the treated group
after applying weights_treated to it}

\item{weighted_outcome_comparison}{the mean outcome for the comparison group
after applying weights_comparison to it}

\item{weighted_outcome_diff}{the difference between the mean outcomes for
the treated and comparison groups after applying weights to them}

\item{alpha_weight}{the contribution of this term to the estimate of alpha}
}
\description{
a class for holding and working with group-time weights
}
