% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{implicit_aipw_weights}
\alias{implicit_aipw_weights}
\title{implicit_aipw_weights}
\usage{
implicit_aipw_weights(
  yname,
  tname,
  idname,
  gname,
  xformula = ~1,
  d_covs_formula = NULL,
  pscore_formula = xformula,
  pscore_d_covs_formula = d_covs_formula,
  extra_balance_vars_formula = NULL,
  extra_balance_d_vars_formula = NULL,
  balance_d_vars_post = TRUE,
  data,
  weightsname = NULL,
  ...
)
}
\arguments{
\item{yname}{the name of the outcome variable}

\item{tname}{the name of the time period variable}

\item{idname}{the name of the id variable}

\item{gname}{the name of the group variable}

\item{xformula}{a formula that includes the covariates that you would like
to include in the AIPW regression.  Time-varying covariates provided in this
argument will be included by taking their level in "base period"
(which in this case is the period right before a particular group
become treated).  Time-invariant covariates can also be included in this
argument.  This is the same way that covariates are included
in \code{did::att_gt}.}

\item{d_covs_formula}{a formula that includes time-varying covariates to
be included in the form $X_t - X_{g-1}$, i.e., their change from the
"base period" to the current period.  By default, covariates are not
included in this way which is the same as in \code{did::att_gt}, but you can
include them by passing a formula here.}

\item{pscore_formula}{a formula that includes the covariates that you would
like to include in the propensity score model.  By default, these covariates
are the same as for the outcome regression which are provided in the earlier
argument \code{xformula}.  This is a natural choice, but this argument allows
these covariates to be different, if desired.

Setting \code{pscore_formula} and \code{xformula} to be different allows for some
important special cases to be handled by this function.  For one, providing
covariates in \code{xformula} while setting \code{pscore_formula=~1} will deliver
deliver "regression adjustment" estimates and decompositions.  For another,
setting \code{pscore_formula} to include covariates while setting \code{xformula=~1}
will deliver "inverse probability weighting" estimates and decompositions.}

\item{pscore_d_covs_formula}{a formula that includes time-varying covariates
to include as $X_t - X_{g-1}$ in the propensity score model.  The default
is to set these to be the same as the covariates in \code{d_covs_formula}.}

\item{extra_balance_vars_formula}{is currently ignored.  This argument
makes the arguments the same as for \code{two_period_aipw_weights}, but currently
covariate balance checking is done in a different way in the multiple
period case by calling \code{aipw_cov_bal} after running this function.}

\item{extra_balance_d_vars_formula}{is currently ignored for the same
reason as \code{extra_balance_vars_formula}.}

\item{balance_d_vars_post}{is currently ignored for the same reason as
\code{extra_balance_vars_formula}.}

\item{data}{a panel data set.  This function is currently only guaranteed
to work with balanced and sorted panel data.  The function tries to
balance and sort, but it is recommended to do this before calling the
function.}

\item{weightsname}{the name of the variable in the data set that contains
the sampling weights.  The default is \code{NULL} which means that all
observations are given equal weight.}

\item{...}{extra arguments, not used here}
}
\value{
a \code{decomposed_aipw} object
}
\description{
a function to compute implicit AIPW weights in a setting with
staggered treatment adoption.
}
