% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{implicit_twfe_weights_gt}
\alias{implicit_twfe_weights_gt}
\title{implicit_twfe_weights_gt}
\usage{
implicit_twfe_weights_gt(
  g,
  tp,
  yname,
  tname,
  idname,
  gname,
  base_period,
  xformula = NULL,
  xname = NULL,
  data,
  weightsname
)
}
\arguments{
\item{g}{the current group for which the weights will be computed}

\item{tp}{the time period for which the weights will be computed}

\item{yname}{The name of the outcome variable}

\item{tname}{The name of the column containing the time periods}

\item{idname}{The individual (cross-sectional unit) id name}

\item{gname}{The name of the variable in \code{data} that
contains the first period when a particular observation is treated.
This should be a positive number for all observations in treated groups.
It defines which "group" a unit belongs to.  It should be 0 for units
in the untreated group.}

\item{base_period}{Whether to use a "varying" base period or a
"universal" base period.  Either choice results in the same
post-treatment estimates of ATT(g,t)'s.  In pre-treatment
periods, using a varying base period amounts to computing a
pseudo-ATT in each treatment period by comparing the change
in outcomes for a particular group relative to its comparison
group in the pre-treatment periods (i.e., in pre-treatment
periods this setting computes changes from period t-1 to period
t, but repeatedly changes the value of t)

A universal base period fixes the base period to always be
(g-anticipation-1).  This does not compute
pseudo-ATT(g,t)'s in pre-treatment periods, but rather
reports average changes in outcomes from period t to
(g-anticipation-1) for a particular group relative to its comparison
group.  This is analogous to what is often reported in event
study regressions.

Using a varying base period results in an estimate of
ATT(g,t) being reported in the period immediately before
treatment.  Using a universal base period normalizes the
estimate in the period right before treatment (or earlier when
the user allows for anticipation) to be equal to 0, but one
extra estimate in an earlier period.}

\item{xformula}{formula for covariates that are included in the
TWFE regression.}

\item{xname}{by default, \code{implicit_twfe_weights_gt} applies the
implicit TWFE regression weights to the outcome.  In
cases where the researcher would like to apply the
implicit regression weights to some other variable, the
name of this variable can be passed in through this
argument.

This is typically used as a way to check how well
the TWFE regression balances levels of time-varying
covariates, time-invariant covariates, or other variables
not included in the model.  This is often useful
as a diagnostic to assess how sensitive the TWFE regression
is to violations of linearity.

@return a gt_weights object
@export}

\item{data}{The name of the data.frame that contains the data}

\item{weightsname}{The name of the column containing the sampling weights.
If not set, all observations have same weight.}
}
\description{
Compute TWFE weights for a particular group
and time period.  This function is developed for the
case where there is a binary treatment and staggered
treatment adoption.  The weights are for the coefficient
on the binary treatment variable and allows for
the regression to include additional time varying
covariates.

This is mainly an internal function and data should already be sorted,
balanced, and time periods and groups should be arranged like 1,2,3,etc.
}
