% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_period_covs.R
\name{log_ratio_sd}
\alias{log_ratio_sd}
\title{log_ratio_sd}
\usage{
log_ratio_sd(x, D, w = rep(1, length(n)))
}
\arguments{
\item{x}{a numeric vector}

\item{D}{a binary vector indicating treated and untreated group}

\item{w}{a numeric vector of weights, default is rep(1, length(n))}
}
\value{
a numeric value containing the log ratio of standard deviations
}
\description{
a helper function to compute the log ratio of standard deviations with treated and untreated group
}
