% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{mp_covariate_bal_summary_helper}
\alias{mp_covariate_bal_summary_helper}
\title{mp_covariate_bal_summary_helper}
\usage{
mp_covariate_bal_summary_helper(decomposed_twfe_obj)
}
\arguments{
\item{decomposed_twfe_obj}{a \code{decomposed_twfe} object that comes from
running \code{implicit_twfe_weights} in a previous step.  The function
can also work when this argument is a \code{decomposed_aipw} coming from
running \code{implicit_aipw_weights}.}
}
\value{
a data frame that contains the balance statistics for each
covariate that was passed into the \code{extra_balance_vars_formula} argument
in either \code{twfe_cov_bal} or \code{aipw_cov_bal}.
}
\description{
a helper function to summarize the balance statistics mainly
to handle some repetitive calculations used across functions.
}
