% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights.R
\name{mp_weights_obj}
\alias{mp_weights_obj}
\title{mp_weights_obj}
\usage{
mp_weights_obj(weights_df)
}
\arguments{
\item{weights_df}{A data.frame containing the following elements:
\itemize{
\item group: group identifier
\item time.period: time period identifier
\item weight: weight for the group-time period
\item attgt: average treatment effect for the group-time period
}}
}
\description{
A class to store weighted averages of group-time average
treatment effects
}
