% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_period_covs.R
\name{pooled_sd}
\alias{pooled_sd}
\title{pooled_sd}
\usage{
pooled_sd(x, D, w = rep(1, length(n)))
}
\arguments{
\item{x}{a numeric vector}

\item{D}{a binary vector indicating treated and untreated group}

\item{w}{a numeric vector of weights, default is rep(1, length(n))}
}
\value{
a numeric value containing the pooled standard deviation
}
\description{
a helper function to compute pooled standard deviations with treated and untreated group
}
