% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{summary.decomposed_aipw}
\alias{summary.decomposed_aipw}
\title{summary.decomposed_aipw}
\usage{
\method{summary}{decomposed_aipw}(object, ...)
}
\arguments{
\item{object}{a \code{decomposed_aipw} object that comes from running
\code{implicit_aipw_weights} or \code{aipw_cov_bal}.}

\item{...}{extra arguments, not used here}
}
\value{
a summary of the decomposition
}
\description{
a function to summarize the results of running \code{implicit_aipw_weights}.
}
