% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{summary.decomposed_twfe}
\alias{summary.decomposed_twfe}
\title{summary.decomposed_twfe}
\usage{
\method{summary}{decomposed_twfe}(object, ...)
}
\arguments{
\item{object}{a decomposed_twfe object}

\item{...}{extra arguments, not used here}
}
\value{
a summary of the decomposition
}
\description{
summarizes a two-way fixed effects regression
decomposition
}
