% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights_mp.R
\name{twfe_cov_bal}
\alias{twfe_cov_bal}
\title{twfe_cov_balance}
\usage{
twfe_cov_bal(decomposed_twfe, extra_balance_vars_formula)
}
\arguments{
\item{decomposed_twfe}{a \code{decomposed_twfe} object that comes from
running \code{implicit_twfe_weights} in a previous step.}

\item{extra_balance_vars_formula}{a formula that includes the variables
that you would like to compute balance statistics for.  The formula
should be in the form of \code{~var1 + var2 + var3 + ...} where \code{var1},
\code{var2}, and \code{var3} can be covariates whose values change over time.
\code{twfe_cov_bal} will average these covariates across all time periods,
and then compute balance statistics for each covariate where the
balance statistics come from the implicit weights inherited from the
TWFE regression.}
}
\value{
a \code{decomposed_twfe} object with the balance statistics added
to the \code{twfe_gt} list element.
}
\description{
A function to compute balance statistics inherited from
running a TWFE regression in a staggered treatment adoption setting.

The function should be called after running \code{implicit_twfe_weights}, and
the it computes balance statistics for the average of each variable
passed in through the \code{extra_balance_vars_formula} argument across
all post-treatment periods.
}
