% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{twfeweights}
\alias{twfeweights}
\title{TWFE Weights}
\description{
A package to compute implicit weights from using two-way
fixed effects (TWFE) regressions in the context of causal inference
with panel data.  Besides implicit TWFE regression weights, the package
also includes weights for other causal inference strategies.
}
