% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_period_covs.R
\name{two_period_covs_obj}
\alias{two_period_covs_obj}
\title{two_period_covs_obj}
\usage{
two_period_covs_obj(
  est,
  weights = NULL,
  dy = NULL,
  D = NULL,
  cov_balance_df = NULL,
  ess = NULL
)
}
\arguments{
\item{est}{the overall estimate, e.g., the alpha coefficient from the two period TWFE regression}

\item{weights}{the implicit weights}

\item{dy}{the outcome variable}

\item{D}{the treatment indicator}

\item{cov_balance_df}{a data frame with the balance statistics for the covariates}

\item{ess}{the effective sample size}
}
\value{
a \code{two_period_covs_obj} object that contains the elements
passed to the function
}
\description{
an object that holds the results of the \code{two_period_reg_weights}
function
}
