% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_period_covs.R
\name{ggtwfeweights.two_period_covs_obj}
\alias{ggtwfeweights.two_period_covs_obj}
\title{ggtwfeweights.two_period_covs_obj}
\usage{
\method{ggtwfeweights}{two_period_covs_obj}(x, ...)
}
\arguments{
\item{x}{the object returned from the \code{two_period_covs} function}

\item{...}{additional arguments, not used here}
}
\value{
a ggplot object
}
\description{
a function to plot the balance statistics from the \code{two_period_covs} function
}
