% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeweights.R
\name{twfe_weights}
\alias{twfe_weights}
\title{twfe_weights}
\usage{
twfe_weights(attgt_res)
}
\arguments{
\item{attgt_res}{Results from did::att_gt function}
}
\value{
a data.frame containing the TWFE weighs
}
\description{
A function to compute TWFE regression weights on ATT(g,t)
in a staggered treatment adoption setting.
}
