% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_period_covs.R
\name{two_period_covs_obj}
\alias{two_period_covs_obj}
\title{two_period_covs_obj}
\format{
A list with the following components:
\item{est}{the two period overall estimate, e.g., the alpha coefficient from the two period TWFE regression}
\item{weights}{the implicit weights}
\item{cov_balance_df}{a data frame with the balance statistics for the covariates}
}
\usage{
two_period_covs_obj(est, weights, dy, D, cov_balance_df)
}
\description{
an object that holds the results of the \code{two_period_covs} function
}
